improv: Include tenant ids in the setup
