+++ packages/rownd-nodejs/src/pluginImplementation.ts
@@ -64,6 +64,7 @@ export function mapRowndUserToSuperTokens(
       thirdPartyUserId: rowndUserData.google_id,
       email: rowndUserData.email,
       isVerified: !!rowndUserVerifiedData.google_id,
+      ...(tenantIds ? { tenantIds } : {}),
     });
   }
 
@@ -78,6 +79,7 @@ export function mapRowndUserToSuperTokens(
       thirdPartyUserId: rowndUserData.apple_id,
       email: rowndUserData.email,
       isVerified: !!rowndUserVerifiedData.apple_id,
+      ...(tenantIds ? { tenantIds } : {}),
     });
   }
 
@@ -86,6 +88,7 @@ export function mapRowndUserToSuperTokens(
       recipeId: "passwordless",
       phoneNumber: rowndUserData.phone_number,
       isVerified: !!rowndUserVerifiedData.phone_number,
+      ...(tenantIds ? { tenantIds } : {}),
     });
   }
 
@@ -94,6 +97,7 @@ export function mapRowndUserToSuperTokens(
       recipeId: "passwordless",
       email: rowndUserData.email,
       isVerified: !!rowndUserVerifiedData.email,
+      ...(tenantIds ? { tenantIds } : {}),
     });
   }
